\name{bhatta_cv}
\alias{bhatta_cv}
\title{Cross-validation approach to select the optimal reduced dimension}
\usage{
  bhatta_cv(x, y, q = NULL, num_folds = 10, ...)
}
\arguments{
  \item{x}{matrix containing the data. The rows are the
  observations, and the columns are the features}

  \item{y}{vector of class labels for each observation}

  \item{q}{vector of the reduced dimensions considered. By
  default, we ... TODO}
}
\description{
  TODO
}

