% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{center_data}
\alias{center_data}
\title{Centers the observations in a matrix by their respective class sample means}
\usage{
center_data(x, y)
}
\arguments{
\item{x}{matrix containing the training data. The rows are the sample
observations, and the columns are the features.}

\item{y}{vector of class labels for each training observation}
}
\value{
matrix with observations centered by its corresponding class sample
mean
}
\description{
Centers the observations in a matrix by their respective class sample means
}
