% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{dmvnorm_diag}
\alias{dmvnorm_diag}
\title{Computes multivariate normal density with a diagonal covariance matrix}
\usage{
dmvnorm_diag(x, mean, sigma)
}
\arguments{
\item{x}{matrix}

\item{mean}{vector of means}

\item{sigma}{vector containing diagonal covariance matrix}
}
\value{
multivariate normal density
}
\description{
Alternative to \code{mvtnorm::dmvnorm}
}
