% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrda.r
\name{plot.hdrda_cv}
\alias{plot.hdrda_cv}
\title{Plots a heatmap of cross-validation error grid for a HDRDA classifier object.}
\usage{
\method{plot}{hdrda_cv}(x, ...)
}
\arguments{
\item{x}{object to plot}

\item{...}{unused}
}
\description{
Uses \code{\link[ggplot2]{ggplot}} to plot a heatmap of the training error
grid.
}

