% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{posterior_probs}
\alias{posterior_probs}
\title{Computes posterior probabilities via Bayes Theorem under normality}
\usage{
posterior_probs(x, means, covs, priors)
}
\arguments{
\item{x}{matrix of observations}

\item{means}{list of means for each class}

\item{covs}{list of covariance matrices for each class}

\item{priors}{list of prior probabilities for each class}
}
\value{
matrix of posterior probabilities for each observation
}
\description{
Computes posterior probabilities via Bayes Theorem under normality
}
