% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqda.r
\name{print.dqda}
\alias{print.dqda}
\title{Outputs the summary for a DQDA classifier object.}
\usage{
\method{print}{dqda}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Summarizes the trained DQDA classifier in a nice manner.
}

