% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrda.r
\name{print.hdrda}
\alias{print.hdrda}
\title{Outputs the summary for a HDRDA classifier object.}
\usage{
\method{print}{hdrda}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Summarizes the trained hdrda classifier in a nice manner.
}
