% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda-diag.r
\name{print.lda_diag}
\alias{print.lda_diag}
\title{Outputs the summary for a DLDA classifier object.}
\usage{
\method{print}{lda_diag}(x, ...)
}
\value{
\code{x} (invisibly).
}
\description{
Summarizes the trained DLDA classifier in a nice manner.
}
\keyword{internal}
