% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda-eigen.r
\name{print.lda_eigen}
\alias{print.lda_eigen}
\title{Outputs the summary for a MDMP classifier object.}
\usage{
\method{print}{lda_eigen}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Summarizes the trained lda_eigen classifier in a nice manner.
}
