% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda-emp-bayes.r
\name{print.lda_emp_bayes}
\alias{print.lda_emp_bayes}
\title{Outputs the summary for a MDEB classifier object.}
\usage{
\method{print}{lda_emp_bayes}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Summarizes the trained lda_emp_bayes classifier in a nice manner.
}
