% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda-emp-bayes-eigen.r
\name{print.lda_emp_bayes_eigen}
\alias{print.lda_emp_bayes_eigen}
\title{Outputs the summary for a MDMEB classifier object.}
\usage{
\method{print}{lda_emp_bayes_eigen}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Summarizes the trained lda_emp_bayes_eigen classifier in a nice manner.
}
