% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda-schafer.r
\name{print.lda_schafer}
\alias{print.lda_schafer}
\title{Outputs the summary for a lda_schafer classifier object.}
\usage{
\method{print}{lda_schafer}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Summarizes the trained lda_schafer classifier in a nice manner.
}
