% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda-shrink-cov.r
\name{print.lda_shrink_cov}
\alias{print.lda_shrink_cov}
\title{Outputs the summary for a SDLDA classifier object.}
\usage{
\method{print}{lda_shrink_cov}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Summarizes the trained SDLDA classifier in a nice manner.
}
