% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda-shrink-mean.r
\name{print.lda_shrink_mean}
\alias{print.lda_shrink_mean}
\title{Outputs the summary for a SmDLDA classifier object.}
\usage{
\method{print}{lda_shrink_mean}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Summarizes the trained SmDLDA classifier in a nice manner.
}
