% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda-thomaz.r
\name{print.lda_thomaz}
\alias{print.lda_thomaz}
\title{Outputs the summary for a lda_thomaz classifier object.}
\usage{
\method{print}{lda_thomaz}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Summarizes the trained lda_thomaz classifier in a nice manner.
}
