% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdeb.r
\name{print.mdeb}
\alias{print.mdeb}
\title{Outputs the summary for a MDEB classifier object.}
\usage{
\method{print}{mdeb}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Summarizes the trained mdeb classifier in a nice manner.
}
