% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdmeb.r
\name{print.mdmeb}
\alias{print.mdmeb}
\title{Outputs the summary for a MDMEB classifier object.}
\usage{
\method{print}{mdmeb}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Summarizes the trained mdmeb classifier in a nice manner.
}
