% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdmp.r
\name{print.mdmp}
\alias{print.mdmp}
\title{Outputs the summary for a MDMP classifier object.}
\usage{
\method{print}{mdmp}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Summarizes the trained mdmp classifier in a nice manner.
}

