% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qda-diag.r
\name{print.qda_diag}
\alias{print.qda_diag}
\title{Outputs the summary for a DQDA classifier object.}
\usage{
\method{print}{qda_diag}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Summarizes the trained DQDA classifier in a nice manner.
}
