% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qda-shrink-cov.r
\name{print.qda_shrink_cov}
\alias{print.qda_shrink_cov}
\title{Outputs the summary for a SDQDA classifier object.}
\usage{
\method{print}{qda_shrink_cov}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Summarizes the trained SDQDA classifier in a nice manner.
}
