% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qda-shrink-mean.r
\name{print.qda_shrink_mean}
\alias{print.qda_shrink_mean}
\title{Outputs the summary for a SmDQDA classifier object.}
\usage{
\method{print}{qda_shrink_mean}(x, ...)
}
\description{
Summarizes the trained SmDQDA classifier in a nice manner.
}
\keyword{internal}
