% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda-high-dim.r
\name{print.rda_high_dim}
\alias{print.rda_high_dim}
\title{Outputs the summary for a HDRDA classifier object.}
\usage{
\method{print}{rda_high_dim}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Summarizes the trained rda_high_dim classifier in a nice manner.
}
