% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdlda.r
\name{print.sdlda}
\alias{print.sdlda}
\title{Outputs the summary for a SDLDA classifier object.}
\usage{
\method{print}{sdlda}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Summarizes the trained SDLDA classifier in a nice manner.
}
