% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdqda.r
\name{print.sdqda}
\alias{print.sdqda}
\title{Outputs the summary for a SDQDA classifier object.}
\usage{
\method{print}{sdqda}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Summarizes the trained SDQDA classifier in a nice manner.
}
