% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smdlda.r
\name{print.smdlda}
\alias{print.smdlda}
\title{Outputs the summary for a SmDLDA classifier object.}
\usage{
\method{print}{smdlda}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Summarizes the trained SmDLDA classifier in a nice manner.
}

