% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smdqda.r
\name{print.smdqda}
\alias{print.smdqda}
\title{Outputs the summary for a SmDQDA classifier object.}
\usage{
\method{print}{smdqda}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Summarizes the trained SmDQDA classifier in a nice manner.
}
