\name{simdiag_transform}
\alias{simdiag_transform}
\title{Transforms a matrix or vector with a simultaneous diagonalizing matrix.}
\usage{
  simdiag_transform(obj, x, q = NULL)
}
\arguments{
  \item{obj}{object of type 'simdiag'}

  \item{x}{matrix or vector}

  \item{q}{integer. (Optional) Dimension of subspace to
  which to project.}
}
\value{
  matrix
}
\description{
  TODO
}

