% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{two_class_sim_data}
\alias{two_class_sim_data}
\title{Example bivariate classification data from caret}
\value{
\item{two_class_sim_data}{a tibble}
}
\description{
Example bivariate classification data from caret
}
\details{
These data were generated using by invoking the `twoClassSim()` 
function in the `caret` package.
}
\examples{
data(two_class_sim_data)
}
\keyword{datasets}
