\name{update_hdrda}
\alias{update_hdrda}
\title{Helper function to update tuning parameters for the HDRDA classifier}
\usage{
  update_hdrda(obj, lambda = 1, gamma = 0)
}
\arguments{
  \item{obj}{a \code{hdrda} object}

  \item{lambda}{a numeric value between 0 and 1,
  inclusively}

  \item{gamma}{a numeric value (nonnegative)}
}
\value{
  a \code{hdrda} object with updated estimates
}
\description{
  TODO
}

