% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda-high-dim.r
\name{update_rda_high_dim}
\alias{update_rda_high_dim}
\title{Helper function to update tuning parameters for the HDRDA classifier}
\usage{
update_rda_high_dim(obj, lambda = 1, gamma = 0)
}
\arguments{
\item{obj}{a \code{rda_high_dim} object}

\item{lambda}{a numeric value between 0 and 1, inclusively}

\item{gamma}{a numeric value (nonnegative)}
}
\value{
a \code{rda_high_dim} object with updated estimates
}
\description{
This function updates some of the quantities in the HDRDA classifier based on
updated values of \code{lambda} and \code{gamma}. The update can greatly
expedite cross-validation to examine a large grid of values for \code{lambda}
and \code{gamma}.
}
