% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-cov.r
\name{cov_list}
\alias{cov_list}
\title{Computes the covariance-matrix maximum likelihood estimators for each class
and returns a list.}
\usage{
cov_list(x, y)
}
\arguments{
\item{x}{data matrix with `n` observations and `p` feature vectors}

\item{y}{class labels for observations (rows) in `x`}
}
\value{
list of the sample covariance matrices of size \eqn{p \times p} for
each class given in `y`.
}
\description{
For a sample matrix, `x`, we compute the MLE for the covariance matrix
for each class given in the vector, `y`.
}
