% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-cov.r
\name{cov_mle}
\alias{cov_mle}
\title{Computes the maximum likelihood estimator for the sample covariance matrix
under the assumption of multivariate normality.}
\usage{
cov_mle(x, diag = FALSE)
}
\arguments{
\item{x}{data matrix with `n` observations and `p` feature vectors}

\item{diag}{logical value. If TRUE, assumes the population covariance matrix
is diagonal. By default, we assume that `diag` is `FALSE`.}
}
\value{
sample covariance matrix of size \eqn{p \times p}. If `diag` is
`TRUE`, then a vector of length `p` is returned instead.
}
\description{
For a sample matrix, `x`, we compute the sample covariance matrix of the
data as the maximum likelihood estimator (MLE) of the population covariance
matrix.
}
\details{
If the `diag` option is set to `TRUE`, then we assume the population
covariance matrix is diagonal, and the MLE is computed under this assumption.
In this case, we return a vector of length `p` instead.
}
