% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{log_determinant}
\alias{log_determinant}
\title{Computes the log determinant of a matrix.}
\usage{
log_determinant(x)
}
\arguments{
\item{x}{matrix}
}
\value{
log determinant of \code{x}
}
\description{
Computes the log determinant of a matrix.
}
