% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda-diag.r
\name{print.lda_diag}
\alias{print.lda_diag}
\title{Outputs the summary for a DLDA classifier object.}
\usage{
\method{print}{lda_diag}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Summarizes the trained DLDA classifier in a nice manner.
}
