% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda-pseudo.r
\name{print.lda_pseudo}
\alias{print.lda_pseudo}
\title{Outputs the summary for a lda_pseudo classifier object.}
\usage{
\method{print}{lda_pseudo}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Summarizes the trained lda_pseudo classifier in a nice manner.
}
