% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qda-diag.r
\name{qda_diag}
\alias{qda_diag}
\alias{qda_diag.default}
\alias{qda_diag.formula}
\alias{predict.qda_diag}
\title{Diagonal Quadratic Discriminant Analysis (DQDA)}
\usage{
qda_diag(x, ...)

\method{qda_diag}{default}(x, y, prior = NULL, ...)

\method{qda_diag}{formula}(formula, data, prior = NULL, ...)

\method{predict}{qda_diag}(object, newdata, type = c("class", "prob", "score"), ...)
}
\arguments{
\item{x}{Matrix or data frame containing the training data. The rows are the 
sample observations, and the columns are the features. Only complete data are 
retained.}

\item{...}{additional arguments (not currently used).}

\item{y}{Vector of class labels for each training observation. Only complete 
data are retained.}

\item{prior}{Vector with prior probabilities for each class. If NULL
(default), then equal probabilities are used. See details.}

\item{formula}{A formula of the form `groups ~ x1 + x2 + ...` That is,
the response is the grouping factor and the right hand side specifies the
(non-factor) discriminators.}

\item{data}{data frame from which variables specified in `formula` are
preferentially to be taken.}

\item{object}{Fitted model object}

\item{newdata}{Matrix or data frame of observations to predict. Each row 
corresponds to a new observation.}

\item{type}{Prediction type: either `"class"`, `"prob"`, or `"score"`.}
}
\value{
`qda_diag` object that contains the trained DQDA classifier
}
\description{
Given a set of training data, this function builds the Diagonal Quadratic
Discriminant Analysis (DQDA) classifier, which is often attributed to Dudoit
et al. (2002). The DQDA classifier belongs to the family of Naive Bayes
classifiers, where the distributions of each class are assumed to be
multivariate normal. Note that the DLDA classifier is a special case of the
DQDA classifier.

The DQDA classifier is a modification to QDA, where the off-diagonal elements
of the pooled sample covariance matrix are set to zero.
}
\details{
The DQDA classifier is a modification to the well-known QDA classifier, where
the off-diagonal elements of each class covariance matrix are assumed
to be zero -- the features are assumed to be uncorrelated. Under multivariate
normality, the assumption uncorrelated features is equivalent to the
assumption of independent features. The feature-independence assumption is a
notable attribute of the Naive Bayes classifier family. The benefit of these
classifiers is that they are fast and have much fewer parameters to estimate,
especially when the number of features is quite large.

The matrix of training observations are given in `x`. The rows of `x`
contain the sample observations, and the columns contain the features for each
training observation.

The vector of class labels given in `y` are coerced to a `factor`.
The length of `y` should match the number of rows in `x`.

An error is thrown if a given class has less than 2 observations because the
variance for each feature within a class cannot be estimated with less than 2
observations.

The vector, `prior`, contains the _a priori_ class membership for
each class. If `prior` is NULL (default), the class membership
probabilities are estimated as the sample proportion of observations belonging
to each class. Otherwise, `prior` should be a vector with the same length
as the number of classes in `y`. The `prior` probabilities should be
nonnegative and sum to one.
}
\examples{
library(modeldata)
data(penguins)
predict_rows <- seq(1, 344, by = 20)
penguins <- penguins[, c("species", "body_mass_g", "flipper_length_mm")]
dqda_out <- qda_diag(species ~ ., data = penguins[-predict_rows, ])
predicted <- predict(dqda_out, penguins[predict_rows, -1], type = "class")

dqda_out2 <- qda_diag(x = penguins[-predict_rows, -1], y = penguins$species[-predict_rows])
predicted2 <- predict(dqda_out2, penguins[predict_rows, -1], type = "class")
all.equal(predicted, predicted2)
}
\references{
Dudoit, S., Fridlyand, J., & Speed, T. P. (2002). "Comparison of
Discrimination Methods for the Classification of Tumors Using Gene Expression
Data," Journal of the American Statistical Association, 97, 457, 77-87.
}
